requirements.txt
flask
Pillow
transformers
torch
numpy
matplotlib
mediapipe